Fixes? + plugin config name changes
For some reason when I run the plugin from plugin.xml, if the name
doesn't start with ATF it fails to load. So therefore the plugin name
(OSGI Bundi name) must remain as ATF_Recorder_Plugin. I suspect this is
just a cacheing issue on my machine. A few changes for getting the RAS
Recorder name on the plugin are also in this commit. Lastly, this commit
has been tested and works on both debug (running from plugin.xml) runs
and actual update site install runs.